database: redefine board type slightly to prevent json and mysql errors
The jest is that I'm not entirely [512]byte works with json-iterator, changing it to strings seems to map it just fine then? Also, having `omitempty` on Algebraic types when they are 0 omits them from the resulting JSON. This caused a problem on the web application when trying to use that to iterate over a list of boards because it didn't exist. This is probably fixable on the web application by just assuming 0 when it is omitted, some more testing will have to be done.